pyproject.toml
[project]
name = "spire-xls-mcp-server"
version = "0.1.1"
description = "MCP server for Excel file manipulation"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "mcp[cli]>=1.2.0",
    "Spire.Xls.Free>=14.12.4"
]
[[project.authors]]
name = "e-iceblue"
email = "sales@e-iceblue.com"

[project.urls]
Homepage = "https://www.e-iceblue.com/"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project.scripts]
spire-xls-mcp-server = "spire_xls_mcp.__main__:main"

[tool.hatch.build.targets.wheel]
packages = ["src/spire_xls_mcp"]

[tool.hatch.build]
packages = ["src/spire_xls_mcp"]